sensashot-arduino/requirements.txt
# Arduino version dependencies
pyserial==3.5
pillow==10.0.0
tkinter  # Usually included with Python
# Optional: pyinstaller==6.0.0 (for creating standalone executables)